+++ ui/components/findings/table/column-findings.tsx
@@ -1,11 +1,16 @@
 "use client";
 
 import { ColumnDef } from "@tanstack/react-table";
+import { Database } from "lucide-react";
 import { useSearchParams } from "next/navigation";
 
 import { DataTableRowDetails } from "@/components/findings/table";
 import { InfoIcon } from "@/components/icons";
-import { DateWithTime, EntityInfoShort } from "@/components/ui/entities";
+import {
+  DateWithTime,
+  EntityInfoShort,
+  SnippetChip,
+} from "@/components/ui/entities";
 import { TriggerSheet } from "@/components/ui/sheet";
 import {
   DataTableColumnHeader,
@@ -63,7 +68,7 @@ const FindingDetailsCell = ({ row }: { row: any }) => {
   };
 
   return (
-    <div className="flex justify-center">
+    <div className="flex max-w-10 justify-center">
       <TriggerSheet
         triggerComponent={<InfoIcon className="text-primary" size={16} />}
         title="Finding Details"
@@ -105,8 +110,10 @@ export const ColumnFindings: ColumnDef<FindingProps>[] = [
       return (
         <div className="relative flex max-w-[410px] flex-row items-center gap-2 3xl:max-w-[660px]">
           <div className="flex flex-row items-center gap-4">
-            {(delta === "new" || delta === "changed") && (
+            {delta === "new" || delta === "changed" ? (
               <DeltaIndicator delta={delta} />
+            ) : (
+              <div className="w-2" />
             )}
             <p className="mr-7 whitespace-normal break-words text-sm">
               {checktitle}
@@ -116,6 +123,21 @@ export const ColumnFindings: ColumnDef<FindingProps>[] = [
             <Muted isMuted={muted} />
           </span>
         </div>
+      );
+    },
+  },
+  {
+    accessorKey: "resourceName",
+    header: "Resource name",
+    cell: ({ row }) => {
+      const resourceName = getResourceData(row, "name");
+
+      return (
+        <SnippetChip
+          value={resourceName as string}
+          formatter={(value: string) => `...${value.slice(-10)}`}
+          icon={<Database size={16} />}
+        />
       );
     },
   },